Move cryptokey extensions to separate service part
